requirements.txt
google-cloud-translate==3.1.0
google-cloud-storage==1.38.0